% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{addSphere}
\alias{addSphere}
\title{Add a sphere(s) to an NGLView object}
\usage{
addSphere(x, position, color = c(1, 0, 0), radius = 0.25)
}
\arguments{
\item{x}{an \code{NGLVieweR} objects, or its proxy}

\item{position}{a numeric of length 3, or a matrix with three columns}

\item{color}{a color}

\item{radius}{the radius}
}
\description{
Uses a buffer to add spheres to an x
Get inspiration from this:
https://x.org/ngl/api/classes/bufferrepresentation.html#dispose
}
