% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR_example.R
\name{NGLVieweR_example}
\alias{NGLVieweR_example}
\title{Run NGLVieweR example Shiny app}
\usage{
NGLVieweR_example(example = "basic")
}
\arguments{
\item{example}{Example type for which to see an example, possible values are:
"basic", "API", "addSelection", "removeSelection", "snapshot",
"updateAnimation", "updateColor", "updateFocus", "updateFullscreen",
"updateRepresentation", "updateSelection", "updateStage", "updateVisibility"
and "updateZoomMove".}
}
\value{
Call to load Shiny example.
}
\description{
Launch an example to demonstrate how to use \code{NGLvieweR} in
Shiny.
}
\examples{

if (interactive()) {

# Basic example
NGLVieweR_example("basic")

# Example with API calls
NGLVieweR_example("API")

# Function specific example
NGLVieweR_example("addSelection")
}

}
