% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR.R, R/api.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{file_ext}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tools}{\code{\link[tools:fileutils]{file_ext}}}
}}

