% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR.R
\name{setFocus}
\alias{setFocus}
\title{Set Focus}
\usage{
setFocus(NGLVieweR, focus = 0)
}
\arguments{
\item{NGLVieweR}{A NGLVieweR object.}

\item{focus}{Set focus between 0 (default) to 100.}
}
\description{
Set Focus
}
\examples{
NGLVieweR("7CID") \%>\%
  addRepresentation("cartoon", param=list(name="cartoon", colorValue="blue")) \%>\%
  setFocus(0)
}
\seealso{
\code{\link[=updateFocus]{updateFocus()}}

Other options: 
\code{\link{setQuality}()},
\code{\link{snapShot}()},
\code{\link{updateFocus}()},
\code{\link{updateFullscreen}()}
}
\concept{options}
