% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR.R
\name{setQuality}
\alias{setQuality}
\title{Set Quality}
\usage{
setQuality(NGLVieweR, quality = "medium")
}
\arguments{
\item{NGLVieweR}{A NGLVieweR object.}

\item{quality}{Set rendering quality. Can be "low", "medium" (default) or "high".}
}
\value{
setQuality parameter in \code{NGLVieweR} \code{htmlwidgets} object.
}
\description{
Set Quality
}
\examples{
NGLVieweR("7CID") \%>\%
  addRepresentation("cartoon", param=list(name="cartoon", colorValue="blue")) \%>\%
  setQuality("medium")
}
\seealso{
Other options: 
\code{\link{setFocus}()},
\code{\link{snapShot}()},
\code{\link{updateFocus}()},
\code{\link{updateFullscreen}()}
}
\concept{options}
