% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR.R
\name{setScale}
\alias{setScale}
\title{Set Scale}
\usage{
setScale(NGLVieweR, scale = 1)
}
\arguments{
\item{NGLVieweR}{A NGLVieweR object.}

\item{scale}{A numeric value indicating the scale factor (default is 1).}
}
\value{
Updated \code{NGLVieweR} object with new scale parameter.
}
\description{
Set the scale factor for the representation
}
\examples{
NGLVieweR("7CID") \%>\%
addRepresentation("cartoon", 
param=list(name="cartoon", colorValue="blue")) \%>\%
setScale(2)
}
\seealso{
\itemize{
\item \code{\link[=zoomMove]{zoomMove()}}
\item \code{\link[=setRotation]{setRotation()}}
\item \code{\link[=setPosition]{setPosition()}}
}

Other transformations: 
\code{\link{setPosition}()},
\code{\link{setRotation}()},
\code{\link{zoomMove}()}
}
\concept{transformations}
