% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR.R
\name{setSpin}
\alias{setSpin}
\title{Set Spin}
\usage{
setSpin(NGLVieweR, spin = TRUE)
}
\arguments{
\item{NGLVieweR}{A NGLVieweR object.}

\item{spin}{If \code{TRUE} (default), start spinning and stop rocking}
}
\value{
setSpin parameter to \code{TRUE} or \code{FALSE} in \code{NGLVieweR} \code{htmlwidgets} object.
}
\description{
Set Spin animation
}
\examples{
NGLVieweR("7CID") \%>\%
 addRepresentation("cartoon", param=list(name="cartoon", colorValue="blue")) \%>\%
 setSpin(TRUE)
}
\seealso{
\itemize{
\item \code{\link[=setRock]{setRock()}}
\item \code{\link[=updateSpin]{updateSpin()}}
\item \code{\link[=setPosition]{setPosition()}}
}

Other animations: 
\code{\link{setRock}()},
\code{\link{updateRock}()},
\code{\link{updateSpin}()},
\code{\link{updateZoomMove}()}
}
\concept{animations}
