% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{updateRock}
\alias{updateRock}
\title{Update Rock}
\usage{
updateRock(NGLVieweR_proxy, rock = TRUE)
}
\arguments{
\item{NGLVieweR_proxy}{A NGLVieweR object.}

\item{rock}{If \code{TRUE} (default), start rocking and stop spinning.}
}
\description{
Start rock animation and stop spinning. Works on an existing NGLVieweR object in Shinymode.
}
\examples{
\dontrun{
NGLVieweR_proxy("structure") \%>\% updateRock(TRUE)
}

if (interactive()) {
library(shiny)

ui = fluidPage(
  titlePanel("Viewer with API inputs"),
  sidebarLayout(
    sidebarPanel(
      radioButtons("animate", label = "Animation",
      choices = c("None", "Spin", "Rock"), selected = "None")
    ),
    mainPanel(
      NGLVieweROutput("structure")
    )
  )
)
server = function(input, output) {
  output$structure <- renderNGLVieweR({
    NGLVieweR("7CID") \%>\%
      addRepresentation("cartoon", param = list(name = "cartoon", color="red"))
  })

  observeEvent(input$animate,{
    if(input$animate == "Rock"){
      NGLVieweR_proxy("structure") \%>\% updateRock(TRUE)
    } else if(input$animate == "Spin") {
      NGLVieweR_proxy("structure") \%>\% updateSpin(TRUE)
    } else{
      NGLVieweR_proxy("structure") \%>\% updateRock(FALSE) \%>\% updateSpin(FALSE)
    }
  })
 }
shinyApp(ui, server)
}
}
\seealso{
\itemize{
\item \code{\link[=setRock]{setRock()}}
\item \code{\link[=NGLVieweR_example]{NGLVieweR_example()}} See example "updateAnimation".
}

Other animations: 
\code{\link{setRock}()},
\code{\link{setSpin}()},
\code{\link{updateSpin}()},
\code{\link{updateZoomMove}()},
\code{\link{zoomMove}()}
}
\concept{animations}
