% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGLVieweR.R
\name{setRock}
\alias{setRock}
\title{Set rock}
\usage{
setRock(NGLVieweR, rock = TRUE)
}
\arguments{
\item{NGLVieweR}{A NGLVieweR object.}

\item{rock}{If \code{TRUE} (default), start rocking and stop spinning.}
}
\description{
Set rock animation
}
\examples{
NGLVieweR("7CID") \%>\%
 addRepresentation("cartoon", param=list(name="cartoon", colorValue="blue")) \%>\%
 setRock(TRUE)
}
\seealso{
\itemize{
\item \code{\link[=setSpin]{setSpin()}}
\item \code{\link[=updateRock]{updateRock()}}
}

Other animations: 
\code{\link{setSpin}()},
\code{\link{updateRock}()},
\code{\link{updateSpin}()},
\code{\link{updateZoomMove}()},
\code{\link{zoomMove}()}
}
\concept{animations}
